# webserver/doc/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "",
        "version": "0.0.2",
        "title": "Linto admin API documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "contact@linto.ai"
        }
    },
    "host": "linto.ai/api",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/api/flow/postbls/application": {
            "post": {
                "tags": [
                    "flow"
                ],
                "summary": "Publish an application workflow on BLS",
                "operationId": "postApplicationFlowOnBLS",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "payload",
                    "in": "body",
                    "description": "static workflow name",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "workflow_name": {
                                "description": "static workflow name",
                                "type": "string"
                            },
                            "workflowTemplate": {
                                "description": "Workflow template name that will be used",
                                "type": "string"
                            },
                            "sttServiceLanguage": {
                                "description": "Language of the STT service that will be used",
                                "type": "string"
                            },
                            "sttService": {
                                "description": "STT service name that will be used",
                                "type": "string"
                            },
                            "tockApplicationName": {
                                "description": "Tock application name that will be used",
                                "type": "string"
                            }
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/server_response"
                        }
                    }
                }
            }
        },
...
        "/api/workflows/application": {
            "get": {
                "tags": [
                    "workflows_applications"
                ],
                "summary": "Get all application workflows",
                "operationId": "GetAllApplicationWorkflows",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/workflows_application"
                            }

                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "workflows_applications"
                ],
                "summary": "Create an application workflows",
                "operationId": "CreateApplicationWorkflow",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "payload",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "workflowName": {
                                "type": "string"
                            },
                            "workflowTemplate": {
                                "type": "string"
                            },
                            "sttServiceLanguage": {
                                "type": "string"
                            },
                            "sttService": {
                                "type": "string"
                            },
                            "tockApplicationName": {
                                "type": "string"
                            }
                        }

                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/workflows_application"
                            }

                        }
                    }
                }
            }
        },
        "/api/workflows/application/{id}": {
            "get": {
                "tags": [
                    "workflows_applications"
                ],
                "summary": "Get an application workflow by its id",
                "operationId": "GetApplicationWorkflowById",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "path",
                    "required": true,
                    "name": "id",
                    "description": "Application workflow id",
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/workflows_application"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "workflows_applications"
                ],
                "summary": "Delete an application workflow by its id",
                "operationId": "DeleteApplicationWorkflow",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "in": "path",
                        "required": true,
                        "name": "workflowId",
                        "description": "Application workflow id",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "workflowName",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/workflows_application"
                        }
                    }
                }
            }
        },
        "/api/workflows/application/{id}/androidusers": {
            "get": {
                "tags": [
                    "workflows_applications"
                ],
                "summary": "Get a user list associated to application {id}",
                "operationId": "GetAndroidUserByApplication",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "path",
                    "required": true,
                    "name": "id",
                    "description": "Application workflow id",
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "_id": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "applications": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "workflows_application": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "flowId": {
                    "type": "string"
                },
                "created_date": {
                    "type": "string",
                    "format": "date-time"
                },
                "updated_date": {
                    "type": "string",
                    "format": "date-time"
                },
                "flow": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string"
                        },
                        "label": {
                            "type": "string"
                        },
                        "nodes": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        },
                        "configs": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
...
        "server_response": {

            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "enum": ["success", "error"],
                    "required": true
                },
                "msg": {
                    "type": "string",
                    "required": true
                },
                "error": {
                    "type": "object",
                    "required": false
                }
            }
        }
    }
}